+++ x/auth/vesting/client/testutil/suite.go
@@ -196,9 +196,7 @@ func (s *IntegrationTestSuite) TestNewMsgCreatePermanentLockedAccountCmd() {
 	// Synchronize height between test runs, to ensure sequence numbers are
 	// properly updated.
 	height, err := s.network.LatestHeight()
-	if err != nil {
-		s.T().Fatalf("Getting initial latest height: %v", err)
-	}
+	s.Require().NoError(err, "Getting initial latest height")
 	s.T().Logf("Initial latest height: %d", height)
 	for name, tc := range testCases {
 		tc := tc
@@ -218,9 +216,7 @@ func (s *IntegrationTestSuite) TestNewMsgCreatePermanentLockedAccountCmd() {
 			}
 		})
 		next, err := s.network.WaitForHeight(height + 1)
-		if err != nil {
-			s.T().Fatalf("Waiting for height %d: %v", height+1, err)
-		}
+		s.Require().NoError(err, "Waiting for height...")
 		height = next
 		s.T().Logf("Height now: %d", height)
 	}